+++ src/main/java/org/study/pizzaservice/repository/InMemOrderRepository.java
@@ -7,11 +7,11 @@ import org.study.pizzaservice.domain.order.Order;
 
 public class InMemOrderRepository implements OrderRepository {
 
-    List<Order> orders = new ArrayList<Order>();
-    
-    public Long saveOrder(Order order) {
-	orders.add(order);
-	return order.getId();
-    }
+	List<Order> orders = new ArrayList<Order>();
+
+	public int saveOrder(Order order) {
+		orders.add(order);
+		return order.getId();
+	}
 
 }
