+++ src/main/java/org/study/pizzaservice/repository/OrderRepository.java
@@ -3,5 +3,5 @@ package org.study.pizzaservice.repository;
 import org.study.pizzaservice.domain.order.Order;
 
 public interface OrderRepository {
-    Long saveOrder(Order order);
+    int saveOrder(Order order);
 }
